+++ controllers/interfaces.go
@@ -3,7 +3,7 @@ package controllers
 import (
 	"context"
 
-	"github.com/nais/unleasherator/pkg/unleash"
+	"github.com/nais/unleasherator/pkg/unleashclient"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -11,5 +11,5 @@ type UnleashInstance interface {
 	IsReady() bool
 	URL() string
 	AdminToken(context.Context, client.Client, string) ([]byte, error)
-	ApiClient(context.Context, client.Client, string) (*unleash.Client, error)
+	ApiClient(context.Context, client.Client, string) (*unleashclient.Client, error)
 }
