+++ pkg/unleashclient/admin_api_test.go
@@ -1,4 +1,4 @@
-package unleash
+package unleashclient
 
 import (
 	"encoding/json"
