+++ pkg/unleashclient/constants.go
@@ -1,0 +1,7 @@
+package unleashclient
+
+const (
+	HealthEndpoint             = "/health"
+	InstanceAdminStatsEndpoint = "/api/admin/instance-admin/statistics"
+	ApiTokensEndpoint          = "/api/admin/api-tokens"
+)
